+++ checker/src/main/java/org/checkerframework/checker/initialization/InitializationParentAnnotatedTypeFactory.java
@@ -716,6 +716,7 @@ public abstract class InitializationParentAnnotatedTypeFactory
             if (elem.getKind() == ElementKind.CONSTRUCTOR) {
                 AnnotatedDeclaredType returnType = (AnnotatedDeclaredType) t.getReturnType();
                 DeclaredType underlyingType = returnType.getUnderlyingType();
+                // TODO: Make sure we check explicit annotations somewhere.
                 returnType.replaceAnnotation(
                         getUnderInitializationAnnotationOfSuperType(underlyingType));
             }